Encodings.cs: Use checked for GetToBase64BufferSize().
And nuint for the interop. Related to #21 GPT-5.4 Pro bullet 2.

+++ src/Geralt/Interop/Interop.Encodings.cs
@@ -15,7 +15,7 @@ internal static partial class Interop
 
         [LibraryImport(DllName)]
         [UnmanagedCallConv(CallConvs = [typeof(CallConvCdecl)])]
-        internal static partial int sodium_base64_encoded_len(nuint binaryLength, int variant);
+        internal static partial nuint sodium_base64_encoded_len(nuint binaryLength, int variant);
 
         [LibraryImport(DllName)]
         [UnmanagedCallConv(CallConvs = [typeof(CallConvCdecl)])]
